VerifyFiles
This script get files from a repository, compare the Left version with the right version using creation date. If the left version is newer it copy the right version to a backup directory, replace the right version by the left version and at after compare all files in the repository the backup folder is compressed and stored.